+++ types/core/warn.d.ts
@@ -1,6 +1,5 @@
 export function warnMsg(code: number, ...args: any[]): void;
 export function errMsg(code: number, ...args: any[]): void;
-export function registerMessages(map: Map<number, (...args: any[]) => string>): void;
 export let devState: {
     devMode: boolean;
 };